[FP-62] Refactor: 피드백 반영
